Fix lint violations surfaced by ruff 0.16.0
CI installs ruff unpinned (ruff>=0.4), and 0.16.0 expanded its default rule
set, turning 20 previously-clean lines into errors across the repo — most of
them pre-existing on main rather than new in this branch.

- UP037: drop quotes from self-referential type annotations (types.py,
  rag/types.py) now that they resolve under `from __future__ import
  annotations`.
- I001: sort import blocks (report/html.py, test_irt.py, test_judge.py).
- RUF100: remove noqa directives for rules that are not enabled.
- RUF046: drop int() around round(), which already returns an int.
- ISC004: parenthesize implicit string concatenation inside collections.

No behaviour change; 64 tests still pass. space/ is left alone: it is outside
the CI lint scope and its main finding (B008 on gr.Progress() defaults) is the
idiom Gradio requires.

+++ src/caliper/adapters/simulate.py
@@ -409,12 +409,12 @@ class SimulatedRAGSubject(ModelAdapter):
         k = self.n_claims
         # Deterministic number of fabricated claims, so faithfulness recovers
         # 1 - hallucination_rate tightly (up to rounding on k claims).
-        n_bad = int(round(k * self.hallucination_rate))
+        n_bad = round(k * self.hallucination_rate)
         bad = set(_rng(self.seed, "rag-halluc", question).permutation(k)[:n_bad].tolist())
         # Of the genuine claims, only a `context_reliance` share is actually
         # read off the retrieved passages; the rest is parametric recall.
         genuine = [j for j in range(k) if j not in bad]
-        n_ctx = int(round(len(genuine) * self.context_reliance))
+        n_ctx = round(len(genuine) * self.context_reliance)
         from_context = set(genuine[:n_ctx])
 
         sentences = []
